% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R, R/pkgname.R
\docType{package}
\name{globalseq}
\alias{globalSeq}
\alias{globalSeq-package}
\alias{globalseq}
\alias{globalseq-package}
\title{globalSeq}
\description{
Testing for association between
RNA-Seq and high-dimensional data

\strong{Background}
\cr Testing for association between RNA-Seq
and other genomic data is challenging
due to high variability of the former
and high dimensionality of the latter.

\strong{Results}
\cr Using the negative binomial distribution and a random effects model,
we developed an omnibus test that overcomes both difficulties.
It may be conceptualised as a test of overall significance
in regression analysis, where the response variable is overdispersed
and the number of explanatory variables exceeds the sample size.

\strong{Conclusions}
\cr The proposed method can detect genetic and epigenetic alterations
that affect gene expression. It can examine complex regulatory
mechanisms of gene expression.

\strong{Reference}
\cr A Rauschenberger, MA Jonker, MA van de Wiel, and RX Menezes.
Testing for association between RNA-Seq
and high-dimensional data. Manuscript in preparation.
}
\seealso{
The following command opens the vignette:
\cr \code{utils::vignette('globalSeq')}
\cr
\cr \code{\link{omnibus}} tests entire covariate sets
\cr \code{\link{proprius}} shows individual contributions
\cr \code{\link{cursus}} analyses the whole genome
}
\keyword{documentation}

