% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{intern.matrix}
\alias{intern.matrix}
\title{Internal function}
\usage{
intern.matrix(Y)
}
\arguments{
\item{Y}{RNA-Seq data\strong{:}
numeric matrix with \code{q} rows (genes)
and \code{n} columns (samples);
or a SummarizedExperiment object}
}
\value{
The function returns a matrix.
}
\description{
Convert RNA-Seq data to a numeric matrix
}
\examples{
# simulate RNA-Seq data
Y <- matrix(rnbinom(30,mu=10,size=1/0.2),nrow=10,ncol=3)
rownames(Y) <- paste("gene",1:nrow(Y),sep="")
colnames(Y) <- paste("cell",1:ncol(Y),sep="")

# create data structure
# Z <- SummarizedExperiment::SummarizedExperiment(
#      S4Vectors::SimpleList(counts=Y))

# conversion to matrix
# all.equal(Y,intern.matrix(Z))

}
\keyword{internal}
