% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palasso}
\alias{palasso}
\title{Pairwise-adaptive lasso}
\usage{
palasso(y, X, oser = FALSE, trial = FALSE, ...)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{X}{covariates\strong{:}
list of matrices with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{oser}{one-standard-error rule\strong{:}
logical (temporary argument)}

\item{trial}{development mode\strong{:}
logical (temporary argument)}

\item{...}{Further arguments for \link[glmnet]{cv.glmnet}
or \link[glmnet]{glmnet}.}
}
\value{
This function returns an object
of type \link[glmnet]{cv.glmnet},
with the additional slot \code{weights}.
}
\description{
The function \code{palasso} cross-validates
the pairwise-adaptive lasso.
Use this regression technique if
the covariates are numerous and occur in pairs.
}
\details{
Let \code{x} denote one entry of the list \code{X}.
See \link[glmnet]{glmnet} for alternative
specifications of \code{y} and \code{x}.
}
\examples{
n <- 100; p <- 200
y <- rbinom(n=n,size=1,prob=0.5)
X <- lapply(1:2,function(x) matrix(rnorm(n*p),nrow=n,ncol=p))
reg <- palasso(y=y,X=X,family="binomial")

}
\keyword{methods}
