% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scales}
\alias{scales}
\title{Weight scales}
\usage{
scales(x, y = NULL, ...)
}
\arguments{
\item{x}{weights \strong{:}
vector of length \eqn{n}}

\item{y}{weights \strong{:}
vector of length \eqn{n},
or \code{NULL} \eqn{(y=1-x)}}

\item{...}{Graphical arguments\strong{:}
line width \code{lwd},
title \code{main},
colours \code{col}}
}
\value{
This function returns a plot.
}
\description{
The function \code{scales} plots weights.
}
\examples{
x <- runif(10)
scales(x)

}
\keyword{methods}
